move: likes, review, profile.jsx -> mypage 폴더
