+++ src/components/common/mypage/Likes.jsx
@@ -1,9 +1,9 @@
 import { useState, useEffect } from 'react';
-import useAuthStore from '../../core/stores/useAuthStore';
+import useAuthStore from '../../../core/stores/useAuthStore';
 import { IoHeartCircleSharp } from 'react-icons/io5';
 import { IoHeartDislikeCircleSharp } from 'react-icons/io5';
 import { HiOutlineDotsVertical } from 'react-icons/hi';
-import { fetchPharmacies } from '../../core/api/pharm';
+import { fetchPharmacies } from '../../../core/api/pharm';
 import { useNavigate } from 'react-router-dom';
 import { useQuery } from '@tanstack/react-query';
 
@@ -96,7 +96,7 @@ const Likes = () => {
 
                   <div className="flex flex-col ">
                     <span className="text-xl font-bold">{pharmacy.name}</span>
-                    <p className="text-base  text-gray-600">{pharmacy.address}</p>
+                    <p className="text-base  text-gray-600 truncate md">{pharmacy.address}</p>
                     <p className="mt-5 text-xl font-bold text-black ">{pharmacy.time}</p>
                   </div>
                 </div>
